# 8112: Flamegraph model creation performance improvements