--- a/core/org.openjdk.jmc.flightrecorder/src/main/java/org/openjdk/jmc/flightrecorder/stacktrace/tree/Node.java
+++ b/core/org.openjdk.jmc.flightrecorder/src/main/java/org/openjdk/jmc/flightrecorder/stacktrace/tree/Node.java
@@ -42,10 +42,12 @@ import java.util.Objects;
  * A node in the graph of aggregated stack traces.
  */
 public final class Node {
+	private static final Object NOT_COMPUTED = new Object();
+
 	/**
 	 * Integer uniquely identifying this node within our data structure.
 	 */
-	private final Integer nodeId;
+	private Object nodeId = NOT_COMPUTED;
 
 	/**
 	 * The frame associated with this node.
@@ -77,7 +79,6 @@ public final class Node {
 	}
 
 	public Node(Node parent, AggregatableFrame frame) {
-		this.nodeId = computeNodeId(parent, frame);
 		this.parent = parent;
 		this.frame = frame;
 		if (frame == null) {
@@ -93,7 +94,10 @@ public final class Node {
 	 * @return the unique identifier associated with this node.
 	 */
 	public Integer getNodeId() {
-		return nodeId;
+		if (nodeId == NOT_COMPUTED) {
+			nodeId = computeNodeId(parent, frame);
+		}
+		return (Integer) nodeId;
 	}
 
 	/**
@@ -155,8 +159,8 @@ public final class Node {
 			return false;
 		Node other = (Node) obj;
 
-		return Objects.equals(nodeId, other.nodeId) && Objects.equals(frame, other.frame) && weight == other.weight
-				&& cumulativeWeight == other.cumulativeWeight;
+		return Objects.equals(getNodeId(), other.getNodeId()) && Objects.equals(frame, other.frame)
+				&& weight == other.weight && cumulativeWeight == other.cumulativeWeight;
 	}
 
 	@Override