--- a/core/org.openjdk.jmc.flightrecorder/src/main/java/org/openjdk/jmc/flightrecorder/stacktrace/tree/StacktraceTreeModel.java
+++ b/core/org.openjdk.jmc.flightrecorder/src/main/java/org/openjdk/jmc/flightrecorder/stacktrace/tree/StacktraceTreeModel.java
@@ -168,7 +168,9 @@ public class StacktraceTreeModel {
 				continue;
 			}
 			IMemberAccessor<IQuantity, IItem> quantityAccessor = getAccessor(iterable, attribute);
-			iterable.forEach((item) -> addItem(item, stacktraceAccessor, quantityAccessor));
+			for (final var item : iterable) {
+				addItem(item, stacktraceAccessor, quantityAccessor);
+			}
 		}
 	}