… core/tests/org.openjdk.jmc.flightrecorder.rules.test/.classpath
… core/tests/org.openjdk.jmc.flightrecorder.test/src/main/java/org/openjdk/jmc/flightrecorder/test/StacktraceTreeModelTest.java