@aptmac 2023-06-27
Just sneaking in a couple of points while this is still fresh, I'll try to take an actual look at this some time soon.

> When loading this file and switching to the Java Application view (with the Flame Graph visualization already in focus) multiple tasks are generated to create a flamegraph for all event types and all threads. I did not investigate why there would be multiple tasks. "multiple" is "at least two" Then without waiting for the graph to load I go on to filter to the thread that I am interested about which generates multiple tasks to generate a flamegraph for all event types in that thread. Lastly I filter to the Method Profiling Samples which creates the last task, which then has to wait for enough of the other tasks to finish that one of the 3 thread in the pool can run it. All in all with this specific jfr file and the current master (everything including commit [5ace151](https://github.com/openjdk/jmc/commit/5ace151b6dc00096b5b3212edfad40e86f8bcf8d)) in this scenario I am waiting about 55 seconds until I see the graph. (just measured with a stopwatch app)

This is would be a great fix, and there's already an outstanding JIRA ticket for something like this at: https://bugs.openjdk.org/browse/JMC-7080

> I was not able to run the ui tests with "mvn verify -P uitests -Dspotbugs.skip=true" maven fails for the module "org.openjdk.jmc.test.jemmy" with the message saying: "[ERROR] Failed to execute goal org.eclipse.tycho:tycho-surefire-plugin:3.0.4:integration-test (default-integration-test) on project org.openjdk.jmc.test.jemmy: Could not find application "org.openjdk.jmc.rcp.application.app" in the test runtime. Make sure that the test runtime includes the bundle which defines this application." This also happens on the master branch. I do not think I touched anything that would not already be failing a unitest, but if I am not mistaken contributors are expected to run them before their pull requests. So I might need some help with that.

What operating system are you running on?

I had encountered this last week (https://github.com/openjdk/jmc/pull/495#issuecomment-1597556556) when looking at the jmc/core refactoring PR, and from what I can tell it relates back to the commit (https://github.com/openjdk/jmc/commit/ae2fbf359aa8f7612a0d3e6f18857e08cfdfc309) to use JDK 17. I was able to reproduce this consistently on Linux, and @RealCLanger encountered this as well on Mac. I was able to run the uitests without issue on my Windows machine though, unless I had been mistakenly using an older commit.

The good (?) news is that there's no flameview specific uitests to worry about verifying here, so running the uitests would be mainly to verify that these changes don't affect any of the other ui pages.