@RealCLanger 2023-06-28
> > I was not able to run the ui tests with "mvn verify -P uitests -Dspotbugs.skip=true" maven fails for the module "org.openjdk.jmc.test.jemmy" with the message saying: "[ERROR] Failed to execute goal org.eclipse.tycho:tycho-surefire-plugin:3.0.4:integration-test (default-integration-test) on project org.openjdk.jmc.test.jemmy: Could not find application "org.openjdk.jmc.rcp.application.app" in the test runtime. Make sure that the test runtime includes the bundle which defines this application." This also happens on the master branch. I do not think I touched anything that would not already be failing a unitest, but if I am not mistaken contributors are expected to run them before their pull requests. So I might need some help with that.
> 
> What operating system are you running on?
> 
> I had encountered this last week ([#495 (comment)](https://github.com/openjdk/jmc/pull/495#issuecomment-1597556556)) when looking at the jmc/core refactoring PR, and from what I can tell it relates back to the commit ([ae2fbf3](https://github.com/openjdk/jmc/commit/ae2fbf359aa8f7612a0d3e6f18857e08cfdfc309)) to use JDK 17. I was able to reproduce this consistently on Linux, and @RealCLanger encountered this as well on Mac. I was able to run the uitests without issue on my Windows machine though, unless I had been mistakenly using an older commit.

Yes, I see this issue with the uitests and I'm looking into resolving it. In fact, that's what's holding me back from integrating #497. I will try to take a look today and then come back. As I see changes in some .classpath files etc., I would like to integrate #497 first and then see this PR getting merged. But since the oca is not yet verified I guess we still have some time anyway. 😉 

Another hint: Please enable GitHub actions for PR verification. I think you have to turn it on here: https://github.com/grandmasterpixel/jmc/actions