@grandmasterpixel 2023-06-28
> But since the oca is not yet verified I guess we still have some time anyway.

Yes you do :) I am currently in the process of figuring out if I am in the "Individual Agreement" case or the "Company Agreement" case.

> What operating system are you running on? [...]

I am on Linux. If the tests work on Windows it sounds like one of those "Windows specific path or line ending is messing with the tools" problems.