@RealCLanger 2023-06-29
I integrated my project cleanup PR, please merge with master now. And I believe that I have found the root cause for the uitests failure. The explanation (workaround: use the clean target) and a fix can be found in #503.