@grandmasterpixel 2023-06-30
merged it and "clean verify" seems to work.
now the ui tests show some other strange behavior even on master which all of the following comments are about.
One problem I could already fix or work around. The tests on linux seem to have issues with desktop scaling. I needed to go back to 100% from 200%. Otherwise it seems that in MCJemmyTestBase.initialize() line 91
`MC.setRecordingAnalysis(false)` the click coordinates do not seem to be computed correctly with 200% scaling. It does not seems to click the checkmark (not in the right place) and thus fails at that point.

I am currently stuck at an issue where some test opens an "import connections" dialog and then inputs something like "7tmp7testExportImport2835912463012025242.xml" where the expected/asserted input is something like "/tmp/testExportImport2835912463012025242.xml". Which looks like the test driver forgot to hold shift when typing the slashes :)

The tests in error for me currently are:
`
ConnectionExportImportTest.testExportImport:93 » TimeoutExpired State 'Waiting...
  ConnectionExportImportTest.testImportNonExistantFile:125 » Jemmy Action 'typin...
  ConnectionExportImportTest.testSetMasterPassword:135 » TimeoutExpired State 'W...
  CustomConnectionsTest.createNewFolder:66 » TimeoutExpired State 'Waiting for 1...
  CustomConnectionsTest.editConnection:122 » TimeoutExpired State 'Waiting for 1...
  CustomConnectionsTest.invalidConnection:113 » TimeoutExpired State 'Waiting fo...
  CustomConnectionsTest.nestedFoldersConnections:79 » TimeoutExpired State 'Wait...
  CustomConnectionsTest.renameConnection:104 » TimeoutExpired State 'Waiting for...
  CustomConnectionsTest.renameFolder:72 » TimeoutExpired State 'Waiting for 1 or...
  CustomConnectionsTest.validConnection:96 » TimeoutExpired State 'Waiting for 1...
`
But I am sure some of that is only because the dialog never closes. I have literally only just now tried that so I am still confident I will find what is going on there :)

On the oca I will hopefully make progress this evening (german time) but I do not really expect that to be resolved even next week.